requirements.txt
amqp==2.3.2
argcomplete==1.8.2
asn1crypto==0.24.0
beautifulsoup4==4.5.3
billiard==3.5.0.4
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cssselect2==0.2.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
defusedxml==0.5.0
dill==0.2.8.2
Django==1.11
django-datatables-view==1.17.0
django-uuidfield==0.5.0
docx==0.2.4
docx2txt==0.6
EbookLib==0.15
en-core-web-sm==2.0.0
enum34==1.1.6
fcm-django==0.2.19
fpdf==1.7.2
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
kiwisolver==1.0.1
kombu==4.2.1
lxml==4.2.5
matplotlib==3.0.2
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.4
numpy==1.15.4
pandas==0.23.4
pdfkit==0.6.1
pdfminer.six==20170720
Pillow==5.3.0
plac==0.9.6
pocketsphinx==0.1.3
preshed==2.0.1
psycopg2-binary==2.7.5
pycparser==2.19
pycryptodome==3.6.6
pyfcm==1.4.5
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyparsing==2.3.0
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.7.5
python-decouple==3.1
python-pptx==0.6.5
pytz==2018.5
regex==2018.1.10
requests==2.20.1
requests-toolbelt==0.8.0
singledispatch==3.4.0.3
six==1.10.0
textract==1.6.1
thinc==6.12.1
tinycss2==0.6.1
toolz==0.9.0
tqdm==4.29.1
ujson==1.35
urllib3==1.24.1
vine==1.1.4
WeasyPrint==43
webencodings==0.5.1
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==1.1.2
xlwt==1.3.0
